requirements.txt
astroid==2.3.3
colorama==0.4.3
dj-database-url==0.5.0
Django==3.2
django-userforeignkey==0.3.0
djangorestframework==3.10.2
django-ckeditor==6.4.2
django-model-utils==4.2.0
django-js-asset==2.0.0
gunicorn==20.0.4
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==9.2.0
psycopg2==2.8.4
pylint==2.4.4
PyPDF2==1.26.0
python-decouple==3.6
pytz==2019.3
reportlab==3.5.32
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.11.1
xhtml2pdf==0.2.3
psycopg2-binary==2.9.4
